Restaurant Database Project
This program utilizes a database to store information such as the menu
items, tables, and customers. A database is used to check if a table is
occupied, how much a bill for a party would cost as well as the most
popular food item for the day. All of these functions are fleshed out in
the java program.